requirements.txt
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.1
cachelib==0.9.0
click==8.1.3
exceptiongroup==1.0.4
execnet==1.9.0
Flask==2.2.2
Flask-Caching==2.0.1
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
packaging==22.0
pluggy==1.0.0
pyrsistent==0.19.2
python-dotenv==0.21.0
pytz==2022.6
redis==4.4.0
six==1.16.0
SQLAlchemy==1.4.45
tomli==2.0.1
Werkzeug==2.2.2
